Move module to subpackage core
