+++ iracema/core/__init__.py
@@ -1,0 +1,4 @@
+# from .audio import Audio
+# from .timeseries import TimeSeries
+# from .point import Point, PointList
+# from .segment import Segment, SegmentList
